requirements.txt
-e .
django
django-configurations
pytest-xdist
